requirements.txt
# Deep Learning
torch==2.10.0
torchvision==0.15.0
torchaudio==2.0.0

# Computer Vision
opencv-python==4.13.0
Pillow==10.0.0

# Data Science
numpy==1.24.3
pandas==3.0.0
scikit-learn==1.3.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Web Application
streamlit==1.56.0

# Utilities
tqdm==4.65.0

# Monitoring (Optional)
tensorboard==2.13.0

# Development (Optional)
jupyter==1.0.0
notebook==7.0.0
